refactor: extract version.kt
