+++ src/main/kotlin/Main.kt
@@ -6,17 +6,14 @@ import group.CommitGroup
 import group.ConventionalCommit
 import kotlinx.serialization.decodeFromString
 import mu.KotlinLogging
-import org.eclipse.jgit.lib.Ref
 import org.eclipse.jgit.revwalk.RevCommit
-import version.SemVer
-import version.VersionCriteria
 import java.nio.file.Path
 import kotlin.io.path.Path
 import kotlin.io.path.div
 import kotlin.io.path.readText
 import kotlin.io.path.writeText
 
-private val logger = KotlinLogging.logger {}
+val logger = KotlinLogging.logger {}
 
 val cwd = Path("/home/scarf/repo/Marisa")
 val changelogPath = cwd / "docs" / "generator"
@@ -28,14 +25,6 @@ val conventionalTitle = Regex("""\w+(\(\w+\))?!?:\s.+""")
 fun RevCommit.isConventional() = conventionalTitle.matches(shortMessage)
 fun RevCommit.parent(): RevCommit = parents.singleOrNull() ?: this
 
-/** @receiver commits from latest to oldest */
-fun List<ConventionalCommit>.version(begin: SemVer, criteria: VersionCriteria) = asReversed()
-    .fold(begin) { acc, commit ->
-        val version = commit.parseVersion(criteria)
-        if (version != null) logger.debug { "$acc + $version -> ${acc + version} :: $commit" }
-        acc + version
-    }
-
 fun Path.toConfig() = Yaml.default.decodeFromString<Config>(readText())
 
 val resources = Path("src/main/resources")
@@ -45,8 +34,6 @@ val sts = (resources / "slaythespire.yaml").toConfig()
 
 val manual = Path("/home/scarf/repo/Marisa/docs/changelog")
 
-fun Ref.toSemVer() = name.substringAfterLast('v').split(".").map { it.toInt() }
-    .let { (major, minor, patch) -> SemVer(major, minor, patch) }
 
 fun main() {
     KGit.open(cwd.toFile()).run {
